backend/requirements.txt
flask
flask-cors
requests
beautifulsoup4
python-dotenv
google-generativeai
pinecone
feedparser
newspaper3k
lxml_html_clean
